refactor
declare peace cheat( original one doesnt work, can be maybe removed again)
- Refactor and moving around
