--- CSharpSourceCode/CharacterDevelopment/CampaignBehavior/DamselEnvoyOfTheLadyPerkBehavior.cs
@@ -1,7 +1,0 @@
-﻿namespace TOR_Core.CampaignMechanics
-{
-    public class DamselEnvoyOfTheLadyPerkBehavior
-    {
-        
-    }
-}
