+++ CSharpSourceCode/Utilities/TORConsoleCommands.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Helpers;
 using TaleWorlds.CampaignSystem;
+using TaleWorlds.CampaignSystem.Actions;
 using TaleWorlds.CampaignSystem.Party;
 using TaleWorlds.Core;
 using TaleWorlds.Library;
@@ -79,6 +80,41 @@ namespace TOR_Core.Utilities
 
 
             return result;
+        }
+        
+        [CommandLineFunctionality.CommandLineArgumentFunction("declare_peace", "tor")]
+        public static string DeclarePeace(List<string> strings)
+        {
+            if (!CampaignCheats.CheckCheatUsage(ref CampaignCheats.ErrorType))
+                return CampaignCheats.ErrorType;
+            string str1 = "campaign.declare_peace [Faction1] | [Faction2]";
+            if (CampaignCheats.CheckParameters(strings, 0) || CampaignCheats.CheckParameters(strings, 1) || CampaignCheats.CheckHelp(strings))
+                return str1;
+            List<string> separatedNames = CampaignCheats.GetSeparatedNames(strings, "|");
+            if (separatedNames.Count != 2)
+                return str1;
+            string kingdom_str1 = separatedNames[0].ToLower().Replace(" ", "");
+            string kingdom_str2 = separatedNames[1].ToLower().Replace(" ", "");
+            Kingdom faction1 = null;
+            Kingdom faction2 =  null;
+            foreach (var kingdom in Campaign.Current.Kingdoms)
+            {
+                if (kingdom_str1 == kingdom.StringId)
+                {
+                    faction1 = kingdom;
+                }
+                if (kingdom_str2 == kingdom.StringId)
+                {
+                    faction2= kingdom;
+                }
+            }
+            
+            if (faction1 != null && faction2 != null)
+            {
+                MakePeaceAction.Apply(faction1, faction2);
+                return "Peace declared between " + (object) faction1.Name + " and " + (object) faction2.Name;
+            }
+            return faction1 == null ? "Faction is not found: " + kingdom_str1 + "\n" + str1 : "Faction is not found: " + kingdom_str2;
         }
